+++ src/dirent.c
@@ -1,6 +1,7 @@
 #include "dirent.h"
 #include "syscall.h"
+#include <stdint.h>
 
 int getdents(int fd, struct dirent *buf, unsigned int count) {
-    return syscall(SYS_GETDENTS, fd, (int)buf, (int)count);
+    return (int)syscall(SYS_GETDENTS, (uintptr_t)fd, (uintptr_t)buf, (uintptr_t)count);
 }
