+++ src/fcntl.c
@@ -1,9 +1,10 @@
 #include "fcntl.h"
 #include "syscall.h"
 #include <stdarg.h>
+#include <stdint.h>
 
 int open(const char *pathname, int flags, ...) {
-    return syscall(SYS_OPEN, (int)pathname, flags, 0);
+    return (int)syscall(SYS_OPEN, (uintptr_t)pathname, (uintptr_t)flags, 0);
 }
 
 int fcntl(int fd, int cmd, ...) {
@@ -11,5 +12,5 @@ int fcntl(int fd, int cmd, ...) {
     va_start(ap, cmd);
     int arg = va_arg(ap, int);
     va_end(ap);
-    return syscall(SYS_FCNTL, fd, cmd, arg);
+    return (int)syscall(SYS_FCNTL, (uintptr_t)fd, (uintptr_t)cmd, (uintptr_t)arg);
 }
