+++ src/shm.c
@@ -1,20 +1,21 @@
 #include "shm.h"
 #include "syscall.h"
+#include <stdint.h>
 
 int shmget(int key, unsigned int size, int flags) {
-    return syscall(SYS_SHMGET, key, (int)size, flags);
+    return (int)syscall(SYS_SHMGET, (uintptr_t)key, (uintptr_t)size, (uintptr_t)flags);
 }
 
 void* shmat(int shmid, const void* shmaddr, int flags) {
-    int ret = syscall(SYS_SHMAT, shmid, (int)shmaddr, flags);
+    intptr_t ret = syscall(SYS_SHMAT, (uintptr_t)shmid, (uintptr_t)shmaddr, (uintptr_t)flags);
     if (ret == -1) return (void*)-1;
-    return (void*)ret;
+    return (void*)(uintptr_t)ret;
 }
 
 int shmdt(const void* shmaddr) {
-    return syscall(SYS_SHMDT, (int)shmaddr, 0, 0);
+    return (int)syscall(SYS_SHMDT, (uintptr_t)shmaddr, 0, 0);
 }
 
 int shmctl(int shmid, int cmd, void* buf) {
-    return syscall(SYS_SHMCTL, shmid, cmd, (int)buf);
+    return (int)syscall(SYS_SHMCTL, (uintptr_t)shmid, (uintptr_t)cmd, (uintptr_t)buf);
 }
