+++ src/socket.c
@@ -1,35 +1,36 @@
 #include "socket.h"
 #include "syscall.h"
+#include <stdint.h>
 
 int socket(int domain, int type, int protocol) {
-    return __syscall3(SYS_SOCKET, domain, type, protocol);
+    return (int)__syscall3(SYS_SOCKET, (uintptr_t)domain, (uintptr_t)type, (uintptr_t)protocol);
 }
 
 int bind(int fd, const struct sockaddr_in *addr, uint32_t addrlen) {
-    return __syscall3(SYS_BIND, fd, (int)addr, (int)addrlen);
+    return (int)__syscall3(SYS_BIND, (uintptr_t)fd, (uintptr_t)addr, (uintptr_t)addrlen);
 }
 
 int connect(int fd, const struct sockaddr_in *addr, uint32_t addrlen) {
-    return __syscall3(SYS_CONNECT, fd, (int)addr, (int)addrlen);
+    return (int)__syscall3(SYS_CONNECT, (uintptr_t)fd, (uintptr_t)addr, (uintptr_t)addrlen);
 }
 
 int listen(int fd, int backlog) {
-    return __syscall2(SYS_LISTEN, fd, backlog);
+    return (int)__syscall2(SYS_LISTEN, (uintptr_t)fd, (uintptr_t)backlog);
 }
 
 int accept(int fd, struct sockaddr_in *peer, uint32_t *addrlen) {
-    return __syscall3(SYS_ACCEPT, fd, (int)peer, (int)addrlen);
+    return (int)__syscall3(SYS_ACCEPT, (uintptr_t)fd, (uintptr_t)peer, (uintptr_t)addrlen);
 }
 
 int send(int fd, const void *buf, uint32_t len, int flags) {
     /* flags not yet honoured by kernel — pass len in edx, flags ignored */
     (void)flags;
-    return __syscall3(SYS_SEND, fd, (int)buf, (int)len);
+    return (int)__syscall3(SYS_SEND, (uintptr_t)fd, (uintptr_t)buf, (uintptr_t)len);
 }
 
 int recv(int fd, void *buf, uint32_t len, int flags) {
     (void)flags;
-    return __syscall3(SYS_RECV, fd, (int)buf, (int)len);
+    return (int)__syscall3(SYS_RECV, (uintptr_t)fd, (uintptr_t)buf, (uintptr_t)len);
 }
 
 int sendto(int fd, const void *buf, uint32_t len, int flags,
@@ -42,7 +43,7 @@ int sendto(int fd, const void *buf, uint32_t len, int flags,
         .dest   = dest,
         .addrlen= addrlen,
     };
-    return __syscall1(SYS_SENDTO, (int)&args);
+    return (int)__syscall1(SYS_SENDTO, (uintptr_t)&args);
 }
 
 int recvfrom(int fd, void *buf, uint32_t len, int flags,
@@ -55,11 +56,11 @@ int recvfrom(int fd, void *buf, uint32_t len, int flags,
         .src     = src,
         .addrlen = addrlen,
     };
-    return __syscall1(SYS_RECVFROM, (int)&args);
+    return (int)__syscall1(SYS_RECVFROM, (uintptr_t)&args);
 }
 
 int shutdown(int fd, int how) {
-    return __syscall2(SYS_SHUTDOWN, fd, how);
+    return (int)__syscall2(SYS_SHUTDOWN, (uintptr_t)fd, (uintptr_t)how);
 }
 
 int setsockopt(int fd, int level, int optname,
@@ -71,7 +72,7 @@ int setsockopt(int fd, int level, int optname,
         .optval  = optval,
         .optlen  = optlen,
     };
-    return __syscall1(SYS_SETSOCKOPT, (int)&args);
+    return (int)__syscall1(SYS_SETSOCKOPT, (uintptr_t)&args);
 }
 
 int getsockopt(int fd, int level, int optname,
@@ -83,5 +84,5 @@ int getsockopt(int fd, int level, int optname,
         .optval  = optval,
         .optlen  = optlen,
     };
-    return __syscall1(SYS_GETSOCKOPT, (int)&args);
+    return (int)__syscall1(SYS_GETSOCKOPT, (uintptr_t)&args);
 }
